# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateFundAmountPositionRequest": {
        "description": "接收基金当前金额和可正可负的持有收益。",
        "properties": {
          "currentValue": {
            "pattern": "^-?\\d+(\\.\\d+)?$",
            "title": "Currentvalue",
            "type": "string"
          },
          "groupId": {
            "format": "uuid",
            "title": "Groupid",
            "type": "string"
          },
          "holdingProfit": {
            "pattern": "^-?\\d+(\\.\\d+)?$",
            "title": "Holdingprofit",
            "type": "string"
          },
          "inputDate": {
            "format": "date",
            "title": "Inputdate",
            "type": "string"
          },
          "inputMode": {
            "const": "FUND_AMOUNT",
            "title": "Inputmode",
            "type": "string"
          },
          "instrumentId": {
            "format": "uuid",
            "title": "Instrumentid",
            "type": "string"
          }
        },
        "required": [
          "inputMode",
          "groupId",
          "instrumentId",
          "inputDate",
          "currentValue",
          "holdingProfit"
        ],
        "title": "CreateFundAmountPositionRequest",
        "type": "object"
      },
      "CreateFundSharesPositionRequest": {
        "description": "接收基金份额及总成本或平均成本二选一输入。",
        "properties": {
          "averageCost": {
            "anyOf": [
              {
                "pattern": "^-?\\d+(\\.\\d+)?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Averagecost"
          },
          "costInputMode": {
            "$ref": "#/components/schemas/CostInputMode"
          },
          "groupId": {
            "format": "uuid",
            "title": "Groupid",
            "type": "string"
          },
          "inputDate": {
            "format": "date",
            "title": "Inputdate",
            "type": "string"
          },
          "inputMode": {
            "const": "FUND_SHARES",
            "title": "Inputmode",
            "type": "string"
          },
          "instrumentId": {
            "format": "uuid",
            "title": "Instrumentid",
            "type": "string"
          },
          "quantity": {
            "pattern": "^-?\\d+(\\.\\d+)?$",
            "title": "Quantity",
            "type": "string"
          },
          "totalCost": {
            "anyOf": [
              {
                "pattern": "^-?\\d+(\\.\\d+)?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Totalcost"
          }
        },
        "required": [
          "inputMode",
          "groupId",
          "instrumentId",
          "inputDate",
          "quantity",
          "costInputMode"
        ],
        "title": "CreateFundSharesPositionRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PositionListResponse": {
        "description": "返回当前用户持仓的稳定分页列表。",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PositionResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "pageSize": {
            "title": "Pagesize",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ],
        "title": "PositionListResponse",
        "type": "object"
      },
...
      "PositionSummaryResponse": {
        "description": "返回组合成本、价格完整性和可选权威估值汇总。",
        "properties": {
          "calculatedAt": {
            "format": "date-time",
            "title": "Calculatedat",
            "type": "string"
          },
          "estimatedFundPositionCount": {
            "title": "Estimatedfundpositioncount",
            "type": "integer"
          },
          "groupId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Groupid"
          },
          "holdingProfit": {
            "anyOf": [
              {
                "pattern": "^-?\\d+(\\.\\d+)?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Holdingprofit"
          },
          "intradayHoldingProfit": {
            "anyOf": [
              {
                "pattern": "^-?\\d+(\\.\\d+)?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intradayholdingprofit"
          },
          "intradayMarketValue": {
            "anyOf": [
              {
                "pattern": "^-?\\d+(\\.\\d+)?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intradaymarketvalue"
          },
          "intradayReturnRate": {
            "anyOf": [
              {
                "pattern": "^-?\\d+(\\.\\d+)?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intradayreturnrate"
          },
          "marketValue": {
            "anyOf": [
              {
                "pattern": "^-?\\d+(\\.\\d+)?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Marketvalue"
          },
          "missingPricePositionIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Missingpricepositionids",
            "type": "array"
          },
          "positionCount": {
            "title": "Positioncount",
            "type": "integer"
          },
          "pricedPositionCount": {
            "title": "Pricedpositioncount",
            "type": "integer"
          },
          "returnRate": {
            "anyOf": [
              {
                "pattern": "^-?\\d+(\\.\\d+)?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Returnrate"
          },
          "stalePositionCount": {
            "title": "Stalepositioncount",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/ValuationStatus"
          },
          "todayProfit": {
            "anyOf": [
              {
                "pattern": "^-?\\d+(\\.\\d+)?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Todayprofit"
          },
          "todayProfitPositionCount": {
            "default": 0,
            "title": "Todayprofitpositioncount",
            "type": "integer"
          },
          "totalCost": {
            "pattern": "^-?\\d+(\\.\\d+)?$",
            "title": "Totalcost",
            "type": "string"
          }
        },
        "required": [
          "groupId",
          "status",
          "positionCount",
          "pricedPositionCount",
          "stalePositionCount",
          "missingPricePositionIds",
          "totalCost",
          "marketValue",
          "holdingProfit",
          "returnRate",
          "intradayMarketValue",
          "intradayHoldingProfit",
          "intradayReturnRate",
          "estimatedFundPositionCount",
          "calculatedAt"
        ],
        "title": "PositionSummaryResponse",
        "type": "object"
      },
...
      "ValuationStatus": {
        "description": "表示组合估值是否完整、陈旧或为空。",
        "enum": [
          "COMPLETE",
          "STALE",
          "INCOMPLETE",
          "EMPTY"
        ],
        "title": "ValuationStatus",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "AI Pick Stock API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/position-summary": {
      "get": {
        "description": "返回全部分组或指定分组的持仓汇总。",
        "operationId": "get_position_summary_api_v1_position_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "groupId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Groupid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Position Summary",
        "tags": [
          "positions"
        ]
      }
    },
    "/api/v1/positions": {
      "get": {
        "description": "分页返回当前用户全部或指定分组的持仓。",
        "operationId": "list_positions_api_v1_positions_get",
        "parameters": [
          {
            "in": "query",
            "name": "groupId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Groupid"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Pagesize",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Positions",
        "tags": [
          "positions"
        ]
      },
      "post": {
        "description": "按录入模式为当前用户分组创建唯一股票或基金持仓。",
        "operationId": "create_position_api_v1_positions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "discriminator": {
                  "mapping": {
                    "FUND_AMOUNT": "#/components/schemas/CreateFundAmountPositionRequest",
                    "FUND_SHARES": "#/components/schemas/CreateFundSharesPositionRequest",
                    "STOCK_SHARES": "#/components/schemas/CreateStockPositionRequest"
                  },
                  "propertyName": "inputMode"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/CreateStockPositionRequest"
                  },
                  {
                    "$ref": "#/components/schemas/CreateFundAmountPositionRequest"
                  },
                  {
                    "$ref": "#/components/schemas/CreateFundSharesPositionRequest"
                  }
                ],
                "title": "Payload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Position",
        "tags": [
          "positions"
        ]
      }
    },
    "/api/v1/positions/{position_id}": {
      "delete": {
        "description": "删除当前用户指定持仓。",
        "operationId": "delete_position_api_v1_positions__position_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "position_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Position Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Position",
        "tags": [
          "positions"
        ]
      },
      "get": {
        "description": "返回当前用户拥有的指定持仓。",
        "operationId": "get_position_api_v1_positions__position_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "position_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Position Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Position",
        "tags": [
          "positions"
        ]
      },
      "patch": {
        "description": "按录入模式和版本部分修改持仓或移动分组。",
        "operationId": "update_position_api_v1_positions__position_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "position_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Position Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "discriminator": {
                  "mapping": {
                    "FUND_AMOUNT": "#/components/schemas/UpdateFundAmountPositionRequest",
                    "FUND_SHARES": "#/components/schemas/UpdateFundSharesPositionRequest",
                    "STOCK_SHARES": "#/components/schemas/UpdateStockPositionRequest"
                  },
                  "propertyName": "inputMode"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/UpdateStockPositionRequest"
                  },
                  {
                    "$ref": "#/components/schemas/UpdateFundAmountPositionRequest"
                  },
                  {
                    "$ref": "#/components/schemas/UpdateFundSharesPositionRequest"
                  }
                ],
                "title": "Payload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Position",
        "tags": [
          "positions"
        ]
      }
    },
...
  }
}